Fix: Prevent double 'posts' directory when blog_dir is 'posts'
When blog_dir is configured as 'posts', the path resolution was incorrectly
creating posts/posts/ instead of posts/.

The Material blog plugin expects posts in {blog_dir}/posts/ by default, but
when blog_dir itself is 'posts', we shouldn't add another 'posts' subdirectory.

This fix checks if blog_path is already 'posts' and uses it directly instead
of appending '/posts'.